+++ src/test/java/nextstep/subway/line/acceptance/LineAcceptanceTest.java
@@ -16,7 +16,7 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 import static nextstep.subway.line.acceptance.LineSteps.*;
-import static nextstep.subway.station.StationSteps.지하철역_등록되어_있음;
+import static nextstep.subway.station.acceptance.StationSteps.지하철역_등록되어_있음;
 import static org.assertj.core.api.Assertions.assertThat;
 
 @DisplayName("지하철 노선 관련 기능")
